+++ src/app/accordion/accordion-examples/accordion-examples.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'it-accordion-examples',
+  templateUrl: './accordion-examples.component.html',
+  styleUrls: ['./accordion-examples.component.scss'],
+})
+export class AccordionExamplesComponent {
+  constructor() {}
+}