+++ src/app/accordion/accordion-group-example/accordion-group-example.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'it-accordion-group-example',
+  templateUrl: './accordion-group-example.component.html',
+  styleUrls: ['./accordion-group-example.component.scss'],
+})
+export class AccordionGroupExampleComponent {}