+++ src/app/app-routing.module.ts
@@ -32,6 +32,8 @@ const routes: Routes = [
       { path: 'button', loadChildren: () => import('src/app/button/button.module').then(m => m.ButtonModule) },
       { path: 'header', loadChildren: () => import('src/app/header/header.module').then(m => m.HeaderModule) },
       { path: 'collapse', loadChildren: () => import('src/app/collapse/collapse.module').then(m => m.CollapseModule) },
+      { path: 'accordion', loadChildren: () => import('src/app/accordion/accordion.module').then(m => m.AccordionModule) },
+
       { path: 'form-input', loadChildren: () => import('src/app/form-input/form-input.module').then(m => m.FormInputModule) },
       { path: 'dropdown', loadChildren: () => import('src/app/dropdown/dropdown.module').then(m => m.DropdownModule) },
       { path: 'back-button', loadChildren: () => import('src/app/back-button/back-button.module').then(m => m.BackButtonModule) },