+++ src/app/app.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 const { version: appVersion } = require('../../package.json');
+import TableOfContent from '../assets/table-of-content.json';
 
 @Component({
   selector: 'it-root',
@@ -7,6 +8,7 @@ const { version: appVersion } = require('../../package.json');
   styleUrls: ['./app.component.scss'],
 })
 export class AppComponent {
+  tableOfContent = (<any>TableOfContent).tableOfContent;
   title = 'design-angular-kit-doc';
   version = appVersion;
 }