+++ src/app/collapse/collapse-index/collapse-index.component.ts
@@ -8,10 +8,8 @@ import Documentation from '../../../assets/documentation.json';
 })
 export class CollapseIndexComponent {
   collapseComponent: any;
-  accordionComponent: any;
 
   constructor() {
     this.collapseComponent = (<any>Documentation).components.find(component => component.name === 'ItCollapseComponent');
-    this.accordionComponent = (<any>Documentation).components.find(component => component.name === 'ItAccordionComponent');
   }
 }