+++ src/app/collapse/collapse.module.ts
@@ -7,17 +7,9 @@ import { CollapseExamplesComponent } from './collapse-examples/collapse-examples
 import { CollapseIndexComponent } from './collapse-index/collapse-index.component';
 import { SharedModule } from '../shared/shared.module';
 import { FormsModule } from '@angular/forms';
-import { CollapseGroupExampleComponent } from './collapse-group-example/collapse-group-example.component';
-import { CollapseNestedExampleComponent } from './collapse-nested-example/collapse-nested-example.component';
 
 @NgModule({
   imports: [CommonModule, FormsModule, SharedModule, CollapseRoutingModule],
-  declarations: [
-    CollapseExampleComponent,
-    CollapseExamplesComponent,
-    CollapseIndexComponent,
-    CollapseGroupExampleComponent,
-    CollapseNestedExampleComponent,
-  ],
+  declarations: [CollapseIndexComponent, CollapseExampleComponent, CollapseExamplesComponent],
 })
 export class CollapseModule {}